Moved the customly generated nodes in VTK*.py into custom_nodes Moved the generated nodes into generated_nodes Updated the relative imports to reflect the directory changes Fixed a few typos in the code The info node now uses format to avoid errors when trying to convert fields not castable to string
