--- VTKReaders.py
@@ -1,3 +1,0 @@
-from .gen_VTKReaders import *
-    
-
