--- VTKSources.py
@@ -1,2 +1,0 @@
-from .gen_VTKSources import *
-
