+++ custom_nodes/VTKSources.py
@@ -1,0 +1,2 @@
+from ..generated_nodes.gen_VTKSources import *
+
