+++ generated_nodes/gen_VTKParametricFunc.py
@@ -1,7 +1,7 @@
 # Generated definitions for VTK class group: ParametricFunc
 # VTK version: 9.0.1
 
-from .core import *    
+from ..core import *    
 TYPENAMES = []
 
 #--------------------------------------------------------------
